src/main/java/mathapp/dto/FunctionDTO.java
package mathapp.dto;

import com.fasterxml.jackson.annotation.JsonProperty;

public class FunctionDTO {
    private Long  functionId;
    private static String type;
    private Double xFrom;
    private Double xTo;
    private Integer count;

    public Long getFunctionId() {
        return functionId;
    }

    public void setFunctionId(Long functionId) {
        this.functionId = functionId;
    }

    public static String getType() {
        return type;
    }

    public void setType(String type) {
        this.type = type;
    }

    @JsonProperty("xFrom")
    public Double getXFrom() {
        return xFrom;
    }

    public void setXFrom(Double xFrom) {
        this.xFrom = xFrom;
    }

    @JsonProperty("xTo")
    public Double getXTo() {
        return xTo;
    }

    public void setXTo(Double xTo) {
        this.xTo = xTo;
    }

    public Integer getCount() {
        return count;
    }

    public void setCount(Integer count) {
        this.count = count;
    }
}